requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
networkx>=2.6.0
seaborn>=0.11.0
jupyter>=1.0.0
notebook>=6.4.0
python-dotenv>=0.19.0
requests>=2.26.0
tqdm>=4.62.0
torch>=2.0.0
openai>=1.99.6
pytest>=8.4.1